refactor(assets): extract version mutations into lib/asset-mutations.ts
Pure relocation of every version/metadata/tag/export mutation out of
asset-service.ts into asset-mutations.ts: addVersion, updateMetadata,
renameTagGlobal/removeTagGlobal/applyTags, promoteVersion, deleteVersion,
addExport, relink, retype, the AssetVersionInput/AssetMetadataInput/
AssetExportInput interfaces, the EXPORT_FORMATS/MAX_EXPORT_DIM consts, and the
mirrorDisplay/removeFileQuietly leaf helpers. All still behind the per-asset
lock + atomic manifest write.

asset-mutations imports getAsset/listAssets/extOf/nowIso from asset-core (never
the barrel), so no cycle. asset-service.ts re-exports the mutation public surface
(consumers/tests unchanged) and now imports addVersion/updateMetadata from
asset-mutations for the upsert path. Dropped the now-unused logger (and const
log), getMimeType, normalizeTags, assetDirAbs, asset-media, manifest, and the
extOf/nowIso asset-core imports. Function bodies byte-for-byte unchanged.

Verified: typecheck/eslint/assets suite 242-0/full suite 5124-0/binary build
(stamps reverted) + boot smoke (assets activates; bakin_assets reindex clean).
